+++ test/integration/zabbix30_postgres/serverspec/zabbix_lwrp_spec.rb
@@ -1,6 +1,25 @@
 require 'spec_helper'
 
-describe file('/etc/apt/sources.list.d/zabbix-official.list') do
+if os[:family] == 'ubuntu'
+  repo_file = '/etc/apt/sources.list.d/zabbix-official.list'
+  postgresql_dir = '/var/lib/postgresql'
+  zabbix_web_package = 'zabbix-frontend-php'
+  if os[:release] == '14.04'
+    php_pgsql_package = 'php5-pgsql'
+    php_zabbix_pool_file = '/etc/php5/fpm/pool.d/zabbix.conf'
+  elsif os[:release] == '16.04'
+    php_pgsql_package = 'php-pgsql'
+    php_zabbix_pool_file = '/etc/php/7.0/fpm/pool.d/zabbix.conf'
+  end
+elsif os[:family] == 'redhat'
+  repo_file = '/etc/yum.repos.d/zabbix.repo'
+  postgresql_dir = '/var/lib/pgsql'
+  zabbix_web_package = 'zabbix-web'
+  php_pgsql_package = 'php-pgsql'
+  php_zabbix_pool_file = '/etc/php-fpm.d/zabbix.conf'
+end
+
+describe file(repo_file) do
   it { should be_file }
 end
 
@@ -39,14 +58,18 @@ describe package('zabbixapi') do
   it { should be_installed.by('gem') }
 end
 
-describe file('/var/lib/postgresql') do
+describe file(postgresql_dir) do
   it { should be_directory }
   it { should be_writable.by_user('postgres') }
-  it { should be_mounted.with(options: { device: '/dev/mapper/shared-zabbix--database' }) }
+  # Check LVM creation on ubuntu only and not Amazon
+  if (os[:family] == 'ubuntu') && !(host_inventory['ec2']['ami-id'])
+    it { should be_mounted.with(options: { device: '/dev/mapper/shared-zabbix--database' }) }
+  end
 end
 
-describe command('pg_lsclusters ') do
-  its(:stdout) { should match '9.4 main    5432 online postgres /var/lib/postgresql/9.4/main' }
+# On CentOS uses old version of postgresql and this command is not exists
+describe command('pg_lsclusters'), if: os[:family] == 'ubuntu' do
+  its(:stdout) { should contain 'main    5432 online postgres' }
 end
 
 describe package('zabbix-server-pgsql') do
@@ -73,15 +96,15 @@ describe port(10_051) do
   it { should be_listening }
 end
 
-describe package('php5-pgsql') do
+describe package(php_pgsql_package) do
   it { should be_installed }
 end
 
-describe package('zabbix-frontend-php') do
+describe package(zabbix_web_package) do
   it { should be_installed }
 end
 
-describe file('/etc/php5/fpm/pool.d/zabbix.conf') do
+describe file(php_zabbix_pool_file) do
   it { should be_file }
   it { should contain 'listen = 127.0.0.1:9200' }
 end
@@ -101,7 +124,7 @@ describe file('/etc/zabbix/web/zabbix.conf.php') do
   it { should contain '10051' }
 end
 
-describe file('/etc/nginx/sites-available/localhost.conf') do
+describe file('/etc/nginx/sites-available/localhost') do
   it { should be_file }
   it { should contain 'listen   80' }
   it { should contain 'server_name  localhost' }