+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file is used to list changes made in each version of zabbix_lwrp.
 
+## 1.2.6
+* Fix creating hosts
+* Rename :modem attribute to :gsm_modem
+
+## 1.2.5
+* Allow to set multiple host groups
+
+## 1.2.4
+* Add trapper_hosts attribute
+
 ## 1.2.3
 * Fix support for Ubuntu 16.04
 * Add hosts sync between zabbix and chef servers